refactor(popup): consolidate type definitions to a single file
